stagefright: Wait longer before force releasing codec
Increase the timeout to 30 secs from 10 in order to give the
components more time to move to idle state, otherwise freeNode
during codec release will also try to move it to idle and bad
things will happen. This is a temporary fix until software
codecs error handling is improved.

Change-Id: I89203bfbce1b81f53ed6e2a33747ec40a786507d

Conflicts:
	media/libstagefright/ACodec.cpp